Bug Fixes
Ensured that the OBS Virtual Camera Is Inactive screen would not appear when the Syphon source had a very slow frame rate.
Ensured that “App Nap” was disabled for Syphon Virtual Camera.
Fixed a bug where a scrambled image would be shown if the Syphon source width was not divisible by 4.
Fixed a few small memory leaks
Added "Download the Latest Version" to the Help Menu